requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
requests
google-generativeai
openai
python-dotenv